backend/requirements.txt
# Core Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# AI Integration
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Optional: For production deployment
gunicorn==21.2.0
python-decouple==3.8

